% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_captchas}
\alias{da_captchas}
\title{Lista de captchas}
\format{
A data frame containing 10 rows and 5 columns

\describe{
\item{site}{Site do captcha}
\item{exemplo}{Exemplo de captcha}
\item{descricao}{Descrição do Captcha}
\item{chutes}{Aceita múltiplos chutes}
\item{n}{Quandidade de casos classificados}
}
}
\usage{
da_captchas
}
\description{
Base com as informações da base
}
\keyword{datasets}
